Assess Consistency of Metrics for Simulated TLS Plots

Plots empirical linear charts of density (N, trees/ha) and basal area (G, \({m}^{2}/ha\)) estimates (derived from simulated TLS plots) as a function of plot size (estimation-size charts) for different plot designs (circular fixed area, k-tree and angle-count), through continuous size increments (radius, k and BAF respectively). Size increments are set at 0.1 m, 1 tree and 0.1 \({m}^{2}/ha\) for fixed area, k-tree and angle-count plot designs, respectively. These size-estimation line charts represent the consistency in predicting the stand variables across different values of radius, k and BAF. Size-estimation charts can be drawn for individual sample plots (including all plots together in the same charts) or for mean values (global mean computed for all the sample plots, or for group means if different strata are considered). Finally, different plot designs can be compared if specified in the arguments, producing one size-estimation chart per variable (N and G).

estimation.plot.size(tree.tls,
                     plot.parameters = data.frame(radius.max = 25,
                                                  k.max = 50,
                                                  BAF.max = 4),
                     dbh.min = 4,
                     average = FALSE, all.plot.designs = FALSE)

Arguments

tree.tls

Data frame with information of trees detected from TLS point cloud data in the same format as tree.detection.single.scan and tree.detection.multi.scan ‘Value’.

plot.parameters

Optional data frame containing parameters for circular fixed area, k-tree and angle-count plot designs. The parameters are as follows:

radius.max: maximum plot radius (m) considered for circular fixed area plots. If the radius.max specified is larger than the farthest tree from the plot centre, the horizontal distance from the farthest tree will be considered the maximum radius. By default, the radius.max will be 25 m.

k.tree.max: maximum number of trees considered for k-tree plots. If k.tree.max specified is larger than the maximum number of trees of the densest plot, this number of trees will be considered the maximum k.tree.max. By default, k.tree.max is 50.

BAF.max: maximum basal area factor (\({m}^{2}/ha\)) considered for angle-count plots. By default, BAF.max is 4.

dbh.min

Optional minimum dbh (cm) considered for detecting trees. By default it will be set at 4 cm.

average

Logical; if TRUE, plot means values and standard deviation of estimations will be represented. By default, it will be set as FALSE.

all.plot.designs

Logical; if TRUE, charts for each plot design are drawn together. By default, it will be set as FALSE.

Details

If there are strata in the tree.tls argument, they will be differentiated in charts with different colours. Strata must be specified in a numeric column named stratum.

The all.plot.designs argument only works for single strata, and therefore if there are additional strata in the tree.tls argument, they will be considered equal.

The outputs of this function are inspired by Fig. 3 of Brunner and Gizachew (2014).

Value

Invisible NULL

References

Brunner, A., & Gizachew, B. (2014). Rapid detection of stand density, tree positions, and tree diameter with a 2D terrestrial laser scanner. European Journal of Forest Research, 133(5), 819-831.

Author

Juan Alberto Molina-Valero, María José Ginzo Villamayor, Manuel Antonio Novo Pérez, Adela Martínez-Calvo, Juan Gabriel Álvarez-González, Fernando Montes and César Pérez-Cruzado.

Note

Mean values are relevant when plots are representing homogenous strata.

Note that this is an option for choosing the best plot design when field data are not available. Otherwise, using correlations, relative.bias and optimize.plot.design will be more desirable for obtaining the best possible plot design.

See also

Examples

# \donttest{

# Loading dataset with trees detected from TLS single-scans

data("Rioja.data")

tree.tls <- Rioja.data$tree.tls


# Without strata and plot parameters by default

estimation.plot.size(tree.tls)
#> Error in .check.class(x = tree.tls, x.class = "data.frame", name = "tree.tls",     n = Inf): 'tree.tls' must be a data.frame.

estimation.plot.size(tree.tls, average = TRUE)
#> Error in .check.class(x = tree.tls, x.class = "data.frame", name = "tree.tls",     n = Inf): 'tree.tls' must be a data.frame.

estimation.plot.size(tree.tls, all.plot.designs = TRUE)
#> Error in .check.class(x = tree.tls, x.class = "data.frame", name = "tree.tls",     n = Inf): 'tree.tls' must be a data.frame.


# Considering two strata and different parameters

# Loading dataset with strata

plot.attributes <- Rioja.data$plot.attributes


# Merging the plot.attributes data set with strata information

tree.tls <- merge(tree.tls, plot.attributes, by = "id")
#> Error in fix.by(by.x, x): 'by' must specify a uniquely valid column


estimation.plot.size(tree.tls,
                     plot.parameters = list(radius.max = 10, k.tree.max = 10, BAF.max = 2))
#> Error in .check.class(x = plot.parameters, x.class = "data.frame", name = "plot.parameters",     n = ifelse(funct %in% c("sim", "est"), 1, Inf)): 'plot.parameters' must be a data.frame.

estimation.plot.size(tree.tls, ,
                     plot.parameters = list(radius.max = 15, k.tree.max = 20, BAF.max = 4),
                     average = TRUE)
#> Error in .check.class(x = plot.parameters, x.class = "data.frame", name = "plot.parameters",     n = ifelse(funct %in% c("sim", "est"), 1, Inf)): 'plot.parameters' must be a data.frame.

estimation.plot.size(tree.tls,
                     plot.parameters = list(radius.max = 20, k.tree.max = 30, BAF.max = 2),
                     all.plot.designs = TRUE)
#> Error in .check.class(x = plot.parameters, x.class = "data.frame", name = "plot.parameters",     n = ifelse(funct %in% c("sim", "est"), 1, Inf)): 'plot.parameters' must be a data.frame.


# }