Calculate dominant diameters and heights for simulations for angle-count plots.

Calculate dominant diameters and heights for simulations for angle-count plots.

Usage

k_tree_cpp(k_seq, radius_seq, k, d, h, num)

Arguments

k_seq

BAF sequence for simulations

radius_seq

BAF sequence for simulations

k

BAF masimum for simulations

d

vector with diameters opf trees

h

vector with heights of trees

num

number of trees considered for calculating diameter and dominant height

Value

Data frame.

Author

Juan Alberto Molina-Valero, María José Ginzo Villamayor, Manuel Antonio Novo Pérez, Adela Martínez-Calvo, Juan Gabriel Álvarez-González, Fernando Montes and César Pérez-Cruzado.